+++ src/main/java/semestrovka/module/repositories/Repository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface Repository<T> {
     List<T> findAll();
-    Optional<T> findById(Long id);
+    Optional<T> findById(int id);
     void save(T entity);
     void update(T entity);
 }
